+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+main:
+  * Replace `StructOp` with `clap`.
+  * Use `LazyLock` instead of mutable static.
+
 v0.2.3 - 2023-01-26:
   * Fix argument parsing with clap 4.